+++ docs/Conventions.md
@@ -18,9 +18,9 @@ Casing style conventions table within this project, with exceptions for predefin
 |-------------------------------|---------------|---|
 | Title Case (with spaces)      | `Short Name.md` | Visible document or Markdown file names where readability matters (e.g. in `docs/`). |
 | UPPER CASE (with spaces)      | `SHORT NAME.md` | Visible document or Markdown file names where readability matters (e.g. in the root `/`). |
-| camelCase                     | `shortName`   | All source file and variable names. |
+| camelCase                     | `shortName`   | All variable names, optionally source file names. |
 | kebab-case                    | `short-name`  | All directory and general file names. |
-| PascalCase                    | `ShortName`   | All class names, interface names (prefixed with `I`), and type names (prefixed with `T`). |
+| PascalCase                    | `ShortName`   | All class names, interface names (prefixed with `I`), and type names (prefixed with `T`). Optionally source file names. |
 | snake_case                    | `short_name`  | _Only used for predefined names._ |
 | UPPER_SNAKE_CASE / SCREAMING_SNAKE_CASE| `SHORT_NAME`  | Optionally used for constant variable names. |
 | kebab-case with caps (non-standard!) | `Short-name`   | _Not recommended — not to be used in source._ |
@@ -32,7 +32,7 @@ For this (Node.js + TypeScript) project, these conventions apply as strictly as 
 |-------------------------------------|------------|------------|
 | Repository and package names        | kebab-case | `yini-parser-typescript`, `yini-parser-ts` |
 | Directory names (within `src/`)     | kebab-case * | `smoke-fixtures` |
-| Source file names                   | camelCase ** | `parseEntry.ts` |
+| Source file names                   | camelCase or PascalCase ** | `parseEntry.ts` or `ErrorHandler.ts` |
 | Script file names                   | kebab-case ** | `clean-ts-js.sh` |
 | Fixture file names (for test data only)| kebab-case ** | `1-basic-key-value.smoke.yini` |
 | Markdown file names                 | Title Case or UPPER CASE ** | `Short Name.md`, `SHORT NAME.md` |
