+++ src/YINIVisitor.ts
@@ -1,11 +1,5 @@
 import assert from 'assert'
 import { isDebug } from './config/env'
-import { extractYiniLine } from './data-extractors/extractSignificantYiniLine'
-import parseBooleanLiteral from './data-extractors/parseBoolean'
-import parseNullLiteral from './data-extractors/parseNull'
-import parseNumberLiteral from './data-extractors/parseNumber'
-import parseSectionHeader from './data-extractors/parseSectionHeader'
-import parseStringLiteral from './data-extractors/parseString'
 import { ErrorDataHandler } from './ErrorDataHandler'
 import {
     Boolean_literalContext,
@@ -29,6 +23,12 @@ import {
     YiniContext,
 } from './grammar/YiniParser.js'
 import YiniParserVisitor from './grammar/YiniParserVisitor'
+import { extractYiniLine } from './parsers/extractSignificantYiniLine'
+import parseBooleanLiteral from './parsers/parseBoolean'
+import parseNullLiteral from './parsers/parseNull'
+import parseNumberLiteral from './parsers/parseNumber'
+import parseSectionHeader from './parsers/parseSectionHeader'
+import parseStringLiteral from './parsers/parseString'
 import {
     IChainContainer,
     ISectionResult,
