+++ src/parsers/parseSectionHeader.ts
